+++ src/test/java/faang/school/accountservice/service/request_task/handler/impl/create_account/CreateAccountTest.java
@@ -57,10 +57,10 @@ class CreateAccountTest {
     private RequestService requestService;
 
     @Mock
-    private CheckAccountsQuantity checkAccountsQuantity;
+    private CheckAccountsQuantityHandler checkAccountsQuantity;
 
     @InjectMocks
-    private CreateAccount createAccount;
+    private CreateAccountHandler createAccount;
 
     @Test
     public void executeTest() throws JsonProcessingException {
